add advanced pcb, customizable hud, fix robofist texture

+++ src/main/java/dev/matthy/betadrive/BatteryPayload.java
@@ -6,10 +6,10 @@ import net.minecraft.network.codec.PacketCodecs;
 import net.minecraft.network.packet.CustomPayload;
 import net.minecraft.util.Identifier;
 
-public record BatteryPayload(float battery) implements CustomPayload {
+public record BatteryPayload(double battery) implements CustomPayload {
     public static final Id<BatteryPayload> ID = new Id<>(Identifier.of("betadrive", "battery"));
     public static final PacketCodec<RegistryByteBuf, BatteryPayload> CODEC = PacketCodec.tuple(
-            PacketCodecs.FLOAT, BatteryPayload::battery,
+            PacketCodecs.DOUBLE, BatteryPayload::battery,
             BatteryPayload::new);
     @Override
     public Id<? extends CustomPayload> getId() {
